+++ src/index.js
@@ -9,7 +9,6 @@ import './js/firebase/firebaseNotifications';
 import './js/firebase/firebase-database';
 import './js/spinner';
 import './js/pagination';
-import './js/modalIneractions';
+import './js/modalInteractions';
 import './js/studentModal';
-import './js/MyLib';
 import './js/upBtn';
