+++ src/js/MyLib.js
@@ -5,54 +5,11 @@ import Notiflix from 'notiflix';
 const WATCHED_KEY = 'Watched_KEY';
 const QUEUE_KEY = 'Queue_KEY';
 
-export const akaLocalStorage = {
-  watched: [],
-  queue: [],
-};
-
 const btnWatched = document.querySelector('.js-btn-watched');
 const btnQueue = document.querySelector('.js-btn-queue');
 
 btnWatched.addEventListener('click', goToWatched);
 btnQueue.addEventListener('click', goToQueue);
-
-export function putWatchedIdtoLocalStorage(event) {
-  const filmId = event.target.dataset.id;
-
-  if (akaLocalStorage.watched.includes(filmId)) {
-    akaLocalStorage.watched.splice(akaLocalStorage.watched.indexOf(filmId), 1);
-
-    const filmSTRING = JSON.stringify(akaLocalStorage.watched);
-    localStorage.setItem(WATCHED_KEY, filmSTRING);
-
-    return;
-  }
-
-  akaLocalStorage.watched.push(filmId);
-
-  const filmSTRING = JSON.stringify(akaLocalStorage.watched);
-
-  localStorage.setItem(WATCHED_KEY, filmSTRING);
-}
-
-export function putQueueIdtoLocalStorage(event) {
-  const filmId = event.target.dataset.id;
-
-  if (akaLocalStorage.queue.includes(filmId)) {
-    akaLocalStorage.queue.splice(akaLocalStorage.queue.indexOf(filmId), 1);
-
-    const filmSTRING = JSON.stringify(akaLocalStorage.queue);
-    localStorage.setItem(QUEUE_KEY, filmSTRING);
-
-    return;
-  }
-
-  akaLocalStorage.queue.push(filmId);
-
-  const filmSTRING = JSON.stringify(akaLocalStorage.queue);
-
-  localStorage.setItem(QUEUE_KEY, filmSTRING);
-}
 
 async function goToWatched() {
   refs.spinner.classList.remove('visually-hidden');
@@ -98,46 +55,31 @@ function renderFilmsMarkup(films) {
   refs.libgallerySet.innerHTML = '';
 
   films
-    .map(
-      ({
-        poster_path,
-        genres,
-        genre_ids,
-        title,
-        original_title,
-        release_date,
-        first_air_date,
-        id,
-      }) => {
-        const poster = poster_path
-          ? `https://image.tmdb.org/t/p/w400${poster_path}`
-          : `https://image.tmdb.org/t/p/w400/yEvumAoCB9Z7o9dAzjxrjcwo2FQ.jpg`;
-        return `<li class="gallery__item" data-id=${id || `No ID`}>
+    .map(({ poster_path, genres, title, original_title, release_date, first_air_date, id }) => {
+      const poster = poster_path
+        ? `https://image.tmdb.org/t/p/w400${poster_path}`
+        : `https://image.tmdb.org/t/p/w400/yEvumAoCB9Z7o9dAzjxrjcwo2FQ.jpg`;
+      return `<li class="gallery__item" data-id=${id || `No ID`}>
                 <div class="films__img">
                     <img src=https://image.tmdb.org/t/p/original${poster} alt="${
-          title || original_title || 'No title'
-        }" loading="lazy" id=${id}>
+        title || original_title || 'No title'
+      }" loading="lazy" id=${id}>
                 </div>
-                <div class="films__description" id=${id}>
-                  <p class="films__title" id=${id}>${
-          title || original_title || 'No title'
-        }</p>
+                <div class="film__description" id=${id}>
+                  <p class="film__title" id=${id}>${title || original_title || 'No title'}</p>
                   <div class="films__meta" id=${id}>
                     <span class="films__genres" id=${id}>${
-          getGenres(genres) || 'No genres info'
-        }</span>
+        getGenres(genres) || 'No genres info'
+      }</span>
                     
                     <span class="films__sep" id=${id}>|</span>
                     <span class="films__data" id=${id}>${
-          new Date(release_date).getFullYear() ||
-          new Date(first_air_date).getFullYear() ||
-          'No info'
-        }</span>
+        new Date(release_date).getFullYear() || new Date(first_air_date).getFullYear() || 'No info'
+      }</span>
                   </div>
                 </div>
             </li>`;
-      }
-    )
+    })
     .forEach(c => refs.libgallerySet.insertAdjacentHTML('beforeend', c));
 }
 
@@ -175,3 +117,5 @@ function getGenres(genres) {
 
   return arr.join(', ');
 }
+
+goToWatched();
