+++ src/myLibrary.html
@@ -12,28 +12,19 @@
       crossorigin="anonymous"
       referrerpolicy="no-referrer"
     />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="32x32"
-      href="./images/favicon.png"
-    />
+    <link rel="icon" type="image/png" sizes="32x32" href="./images/favicon.png" />
     <script type="module" src="./index.js"></script>
   </head>
   <body>
     <include src="./partials/headerML.html"></include>
     <include src="./partials/library.html"> </include>
-
     <include src="./partials/pagination.html"></include>
     <include src="./partials/upBtn.html"></include>
-    <!-- <include src="./partials/genres-switcher.html"></include>
-    <include src="./partials/main-container.html"></include> -->
-
     <include src="./partials/footer.html"></include>
     <include src="./partials/auth-modal.html"></include>
-    <script type="module" src="./js/MyLib.js"></script>
     <include src="./partials/students-modal.html"></include>
     <include src="./partials/modal.html"></include>
 
+    <script type="module" src="./js/MyLib.js"></script>
   </body>
 </html>
