+++ src/sass/components/_body.scss
@@ -1,19 +1,14 @@
 html,
 body {
-  // height: 100%;
   line-height: 1;
   font-family: 'Roboto', sans-serif;
   font-weight: 500;
   margin: 0;
   padding: 0;
-  // overflow-x: hidden;
   scroll-behavior: smooth;
-  overflow-y: scroll;
   background-color: $background-main-color;
 }
 
-
-.stop-scroll{
+.stop-scroll {
   overflow: hidden;
-   
 }
